refactoring endpoints
