+++ __tests__/api/offerings-test.js
@@ -1,14 +1,8 @@
 import axios from 'axios'
 import MockAdapter from 'axios-mock-adapter'
-import {
-  ENDPOINTS as AUTH_ENDPOINTS,
-  getUserMetadata,
-  setAuthToken,
-  signOutUser,
-} from '../../src/api/auth'
+import { getUserMetadata, setAuthToken, signOutUser } from '../../src/api/auth'
 import { format } from '../../src/utils/string'
 import {
-  ENDPOINTS as OFFERING_ENDPOINTS,
   getOfferingData,
   getStarredOfferings,
   getStarredOfferingsData,
@@ -19,13 +13,15 @@ import {
   OFFERINGS_RESPONSE_2,
   STARRED_OFFERINGS_RESPONSE,
 } from '../mock_responses/mock-offerings-response'
+
+import { ENDPOINTS } from '../../src/api/api-requests'
 import { METADATA_RESPONSE } from '../mock_responses/mock-auth-response'
 
 const MOCK_AUTH_TOKEN = 'a'
 const mock = new MockAdapter(axios)
 describe('Get starred offerings', () => {
   beforeEach(() => {
-    mock.onGet(`${AUTH_ENDPOINTS.USER_METADATA}`).reply(HTTP_STATUS_CODES.OK, METADATA_RESPONSE)
+    mock.onGet(`${ENDPOINTS.USER_METADATA}`).reply(HTTP_STATUS_CODES.OK, METADATA_RESPONSE)
   })
 
   afterEach(async () => {
@@ -48,7 +44,7 @@ describe('Get starred offerings', () => {
 
 describe('Get offerings data', () => {
   beforeEach(() => {
-    mock.onGet(`${AUTH_ENDPOINTS.USER_METADATA}`).reply(HTTP_STATUS_CODES.OK, METADATA_RESPONSE)
+    mock.onGet(`${ENDPOINTS.USER_METADATA}`).reply(HTTP_STATUS_CODES.OK, METADATA_RESPONSE)
   })
 
   afterEach(async () => {
@@ -59,7 +55,7 @@ describe('Get offerings data', () => {
   test('when authenticated', async () => {
     const offeringId = 'ac5b1727-629c-443b-8c1a-cc1bd541af6a'
     mock
-      .onGet(`${format(OFFERING_ENDPOINTS.OFFERING, offeringId)}`)
+      .onGet(`${format(ENDPOINTS.OFFERING, offeringId)}`)
       .reply(HTTP_STATUS_CODES.OK, OFFERINGS_RESPONSE_1)
     setAuthToken(MOCK_AUTH_TOKEN)
 
@@ -69,8 +65,8 @@ describe('Get offerings data', () => {
 
   test('with network error', async () => {
     const offeringId = 'ac5b1727-629c-443b-8c1a-cc1bd541af6a'
-    mock.onGet(`${format(OFFERING_ENDPOINTS.OFFERING, offeringId)}`).networkError()
-    mock.onGet(`${OFFERING_ENDPOINTS.OFFERING}`).networkError()
+    mock.onGet(`${format(ENDPOINTS.OFFERING, offeringId)}`).networkError()
+    mock.onGet(`${ENDPOINTS.OFFERING}`).networkError()
 
     const offeringData = await getOfferingData(offeringId)
     expect(offeringData).toBe(null)
@@ -79,7 +75,7 @@ describe('Get offerings data', () => {
 
 describe('Get starred offerings data', () => {
   beforeEach(() => {
-    mock.onGet(`${AUTH_ENDPOINTS.USER_METADATA}`).reply(HTTP_STATUS_CODES.OK, METADATA_RESPONSE)
+    mock.onGet(`${ENDPOINTS.USER_METADATA}`).reply(HTTP_STATUS_CODES.OK, METADATA_RESPONSE)
   })
 
   afterEach(async () => {
@@ -93,10 +89,10 @@ describe('Get starred offerings data', () => {
       '2c7a83cc-e2f3-493a-ae65-33f9c998e8ed',
     ]
     mock
-      .onGet(`${format(OFFERING_ENDPOINTS.OFFERING, offerings[0])}`)
+      .onGet(`${format(ENDPOINTS.OFFERING, offerings[0])}`)
       .reply(HTTP_STATUS_CODES.OK, OFFERINGS_RESPONSE_1)
     mock
-      .onGet(`${format(OFFERING_ENDPOINTS.OFFERING, offerings[1])}`)
+      .onGet(`${format(ENDPOINTS.OFFERING, offerings[1])}`)
       .reply(HTTP_STATUS_CODES.OK, OFFERINGS_RESPONSE_2)
     setAuthToken(MOCK_AUTH_TOKEN)
     await getUserMetadata()
@@ -106,13 +102,13 @@ describe('Get starred offerings data', () => {
   })
 
   test('with network error', async () => {
-    mock.onGet(`${OFFERING_ENDPOINTS.OFFERING}`).networkError()
+    mock.onGet(`${ENDPOINTS.OFFERING}`).networkError()
     const offerings = [
       'ac5b1727-629c-443b-8c1a-cc1bd541af6a',
       '2c7a83cc-e2f3-493a-ae65-33f9c998e8ed',
     ]
-    mock.onGet(`${format(OFFERING_ENDPOINTS.OFFERING, offerings[0])}`).networkError()
-    mock.onGet(`${format(OFFERING_ENDPOINTS.OFFERING, offerings[1])}`).networkError()
+    mock.onGet(`${format(ENDPOINTS.OFFERING, offerings[0])}`).networkError()
+    mock.onGet(`${format(ENDPOINTS.OFFERING, offerings[1])}`).networkError()
     setAuthToken(MOCK_AUTH_TOKEN)
     await getUserMetadata()
