+++ __tests__/frontend/CourseListContainer-test.js
@@ -3,7 +3,7 @@ import React from 'react'
 import MockAdapter from 'axios-mock-adapter'
 import { fireEvent, render, waitFor } from '@testing-library/react-native'
 import { format } from '../../src/utils/string'
-import { ENDPOINTS } from '../../src/api/universities'
+import { ENDPOINTS } from '../../src/api/api-requests'
 import { HTTP_STATUS_CODES } from '../../src/api'
 import { COURSES_RESPONSE } from '../mock_responses/mock-course-response'
 import CourseListContainer from '../../src/containers/CourseListContainer/CourseListContainer'
