+++ __tests__/frontend/DepartmentListContainer-test.js
@@ -3,7 +3,7 @@ import React from 'react'
 import MockAdapter from 'axios-mock-adapter'
 import { fireEvent, render, waitFor } from '@testing-library/react-native'
 import { format } from '../../src/utils/string'
-import { ENDPOINTS } from '../../src/api/universities'
+import { ENDPOINTS } from '../../src/api/api-requests'
 import { HTTP_STATUS_CODES } from '../../src/api'
 import { DEPARTMENTS_RESPONSE } from '../mock_responses/mock-department-response'
 import DepartmentListContainer from '../../src/containers/DepartmentListContainer/DepartmentListContainer'
