+++ __tests__/frontend/PlaylistContainer-test.js
@@ -2,7 +2,7 @@ import axios from 'axios'
 import React from 'react'
 import MockAdapter from 'axios-mock-adapter'
 import { render, waitFor, fireEvent } from '@testing-library/react-native'
-import { ENDPOINTS } from '../../src/api/playlists'
+import { ENDPOINTS } from '../../src/api/api-requests'
 import { HTTP_STATUS_CODES } from '../../src/api'
 import { VIDEOS_BY_PLAYLIST_RESPONSE } from '../mock_responses/mock-playlists-response'
 import { format } from '../../src/utils/string'
