+++ __tests__/frontend/UniversityListContainer-test.js
@@ -2,7 +2,7 @@ import axios from 'axios'
 import React from 'react'
 import MockAdapter from 'axios-mock-adapter'
 import { fireEvent, render, waitFor } from '@testing-library/react-native'
-import { ENDPOINTS } from '../../src/api/universities'
+import { ENDPOINTS } from '../../src/api/api-requests'
 import { HTTP_STATUS_CODES } from '../../src/api'
 import { UNIVERSITY_RESPONSE } from '../mock_responses/mock-university-response'
 import UniversityListContainer from '../../src/containers/UniversityListContainer/UniversityListContainer'
