+++ src/api/playlists.js
@@ -1,11 +1,5 @@
-import { API_BASE_URL } from '../constants'
 import { format } from '../utils/string'
-import { apiCall } from './api-requests'
-
-export const ENDPOINTS = {
-  PLAYLISTS_BY_OFFERING: `${API_BASE_URL}Playlists/ByOffering/{0}`,
-  VIDEOS_BY_PLAYLIST: `${API_BASE_URL}Playlists/{0}`,
-}
+import { apiCall, ENDPOINTS } from './api-requests'
 
 export const getPlaylistsByOffering = async (offeringId) => {
   const url = format(ENDPOINTS.PLAYLISTS_BY_OFFERING, offeringId)
